# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/broker/projects": {
      "get": {
        "operationId": "BrokerMarketplaceController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BrokerProjectResponseDto" }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Broker Marketplace"]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "ProjectsController_list_v1",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Project, developer or location search",
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "name": "provinceId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "districtId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "developerOrganizationId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": { "example": "6000000.0000", "type": "string" }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": { "example": "12000000.0000", "type": "string" }
          },
          {
            "name": "roomType",
            "required": false,
            "in": "query",
            "schema": { "example": "2+1", "type": "string" }
          },
          {
            "name": "minNetArea",
            "required": false,
            "in": "query",
            "description": "Minimum net area of an available unit",
            "schema": { "example": "80.00", "type": "string" }
          },
          {
            "name": "maxDownPaymentPercent",
            "required": false,
            "in": "query",
            "description": "Maximum down-payment percentage",
            "schema": { "example": "30.00", "type": "string" }
          },
          {
            "name": "maxMonthlyPayment",
            "required": false,
            "in": "query",
            "description": "Maximum monthly payment",
            "schema": { "example": "300000.0000", "type": "string" }
          },
          {
            "name": "deliveryReady",
            "required": false,
            "in": "query",
            "description": "Only ready or only not-ready projects",
            "schema": { "type": "boolean" }
          },
          {
            "name": "deliveryBefore",
            "required": false,
            "in": "query",
            "description": "Delivery on or before this date",
            "schema": { "format": "date", "type": "string" }
          },
          {
            "name": "bounds",
            "required": false,
            "in": "query",
            "description": "west,south,east,north decimal bounds",
            "schema": { "type": "string" }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "NEWEST",
              "type": "string",
              "enum": ["NEWEST", "PRICE_ASC", "PRICE_DESC", "DELIVERY_ASC"]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from pageInfo.nextCursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectListResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Projects"]
      }
    },
    "/api/v1/projects/{idOrSlug}": {
      "get": {
        "operationId": "ProjectsController_get_v1",
        "parameters": [
          { "name": "idOrSlug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectDetailResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "tags": ["Projects"]
      }
    },
    "/api/v1/developer/projects": {
      "get": {
        "operationId": "DeveloperProjectsController_list_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": { "type": "string", "enum": ["DRAFT", "IN_REVIEW", "PUBLISHED", "ARCHIVED"] }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "allOf": [{ "$ref": "#/components/schemas/Object" }]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperProjectListResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Projects"]
      },
      "post": {
        "operationId": "DeveloperProjectsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectResponseDto" }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Projects"]
      }
    },
...
    "/api/v1/projects/{idOrSlug}/views": {
      "post": {
        "operationId": "ProjectViewsController_record_v1",
        "parameters": [
          { "name": "idOrSlug", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecordProjectViewDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectViewResponseDto" }
              }
            }
          }
        },
        "tags": ["Project Analytics"]
      }
    },
    "/api/v1/media/{mediaId}": {
      "get": {
        "operationId": "PublicMediaController_get_v1",
        "parameters": [
          { "name": "mediaId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Project Media"]
      }
    },
    "/api/v1/developer/organizations": {
      "get": {
        "operationId": "DeveloperGrowthController_organizations_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/OrganizationScopeResponseDto" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      }
    },
    "/api/v1/developer/analytics": {
      "get": {
        "operationId": "DeveloperGrowthController_analytics_v1",
        "parameters": [
          {
            "name": "range",
            "required": false,
            "in": "query",
            "schema": { "default": 7, "type": "number", "enum": [7, 30] }
          },
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperAnalyticsResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      }
    },
    "/api/v1/developer/settings": {
      "get": {
        "operationId": "DeveloperGrowthController_settings_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperSettingsResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      },
      "patch": {
        "operationId": "DeveloperGrowthController_updateSettings_v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateDeveloperSettingsDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperSettingsResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      }
    },
    "/api/v1/developer/projects/{projectId}/media": {
      "get": {
        "operationId": "DeveloperGrowthController_media_v1",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/DeveloperMediaResponseDto" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      },
      "post": {
        "operationId": "DeveloperGrowthController_upload_v1",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "kind": { "type": "string", "enum": ["IMAGE"], "default": "IMAGE" },
                  "altText": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperMediaResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      }
    },
    "/api/v1/developer/projects/{projectId}/media/{mediaId}": {
      "patch": {
        "operationId": "DeveloperGrowthController_updateMedia_v1",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "mediaId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateMediaDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperMediaResponseDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      },
      "delete": {
        "operationId": "DeveloperGrowthController_deleteMedia_v1",
        "parameters": [
          { "name": "projectId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "mediaId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "version", "required": true, "in": "query", "schema": { "type": "number" } }
        ],
        "responses": { "204": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Developer Growth"]
      }
    },
...
  },
  "info": {
    "title": "PLANDA Product API",
    "description": "Shared API for web, iOS, Android, developer, broker and admin clients",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Validation failed" },
          "details": { "type": "array", "items": { "type": "string" } },
          "requestId": { "type": "string", "example": "4f12bc5f-1a5e-4d14-bf14-6c7bb19ea413" }
        },
        "required": ["statusCode", "code", "message", "requestId"]
      },
...
      "Object": { "type": "object", "properties": {} },
...
      "BrokerProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "developerName": { "type": "string" },
          "publicStartingPrice": { "type": "string" },
          "brokerPrice": { "type": "object", "nullable": true },
          "commissionPercent": { "type": "object", "nullable": true },
          "reservationHours": { "type": "object", "nullable": true },
          "salesContact": { "type": "object", "nullable": true },
          "units": { "type": "array", "items": { "type": "object" } },
          "materialCount": { "type": "number" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "slug",
          "name",
          "developerName",
          "publicStartingPrice",
          "units",
          "materialCount",
          "updatedAt"
        ]
      },
...
      "RecordProjectViewDto": {
        "type": "object",
        "properties": { "sessionId": { "type": "string", "minLength": 16, "maxLength": 200 } },
        "required": ["sessionId"]
      },
...
    }
  }
}